+++ web/assets/js/Control.Geocoder.js
@@ -1,7 +1,7 @@
-(function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
+(function(f){if(typeof exports==="object"&&typeof module!=="undefined"){module.exports=f()}else if(typeof define==="function"&&define.amd){define([],f)}else{var g;if(typeof window!=="undefined"){g=window}else if(typeof global!=="undefined"){g=global}else if(typeof self!=="undefined"){g=self}else{g=this}g.leafletControlGeocoder = f()}})(function(){var define,module,exports;return (function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Nominatim = require('./geocoders/nominatim')["class"];
+	Nominatim = _dereq_('./geocoders/nominatim')["class"];
 
 module.exports = {
 	"class": L.Control.extend({
@@ -289,10 +289,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"./geocoders/nominatim":9}],2:[function(require,module,exports){
+},{"./geocoders/nominatim":9}],2:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -376,10 +376,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],3:[function(require,module,exports){
+},{"../util":13}],3:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -433,10 +433,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],4:[function(require,module,exports){
+},{"../util":13}],4:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -526,10 +526,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],5:[function(require,module,exports){
+},{"../util":13}],5:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-    Util = require('../util');
+    Util = _dereq_('../util');
 
 module.exports = {
     "class": L.Class.extend({
@@ -600,10 +600,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],6:[function(require,module,exports){
+},{"../util":13}],6:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -692,10 +692,10 @@ module.exports = {
 
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],7:[function(require,module,exports){
+},{"../util":13}],7:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -781,10 +781,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],8:[function(require,module,exports){
+},{"../util":13}],8:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -870,10 +870,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],9:[function(require,module,exports){
+},{"../util":13}],9:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -968,10 +968,10 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],10:[function(require,module,exports){
+},{"../util":13}],10:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -1071,10 +1071,10 @@ module.exports = {
 
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],11:[function(require,module,exports){
+},{"../util":13}],11:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Util = require('../util');
+	Util = _dereq_('../util');
 
 module.exports = {
 	"class": L.Class.extend({
@@ -1137,20 +1137,20 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"../util":13}],12:[function(require,module,exports){
+},{"../util":13}],12:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
-	Control = require('./control'),
-	Nominatim = require('./geocoders/nominatim'),
-	Bing = require('./geocoders/bing'),
-	MapQuest = require('./geocoders/mapquest'),
-	Mapbox = require('./geocoders/mapbox'),
-	What3Words = require('./geocoders/what3words'),
-	Google = require('./geocoders/google'),
-	Photon = require('./geocoders/photon'),
-	Mapzen = require('./geocoders/mapzen'),
-	ArcGis = require('./geocoders/arcgis'),
-	HERE = require('./geocoders/here');
+	Control = _dereq_('./control'),
+	Nominatim = _dereq_('./geocoders/nominatim'),
+	Bing = _dereq_('./geocoders/bing'),
+	MapQuest = _dereq_('./geocoders/mapquest'),
+	Mapbox = _dereq_('./geocoders/mapbox'),
+	What3Words = _dereq_('./geocoders/what3words'),
+	Google = _dereq_('./geocoders/google'),
+	Photon = _dereq_('./geocoders/photon'),
+	Mapzen = _dereq_('./geocoders/mapzen'),
+	ArcGis = _dereq_('./geocoders/arcgis'),
+	HERE = _dereq_('./geocoders/here');
 
 module.exports = L.Util.extend(Control["class"], {
 	Nominatim: Nominatim["class"],
@@ -1181,7 +1181,7 @@ L.Util.extend(L.Control, {
 });
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{"./control":1,"./geocoders/arcgis":2,"./geocoders/bing":3,"./geocoders/google":4,"./geocoders/here":5,"./geocoders/mapbox":6,"./geocoders/mapquest":7,"./geocoders/mapzen":8,"./geocoders/nominatim":9,"./geocoders/photon":10,"./geocoders/what3words":11}],13:[function(require,module,exports){
+},{"./control":1,"./geocoders/arcgis":2,"./geocoders/bing":3,"./geocoders/google":4,"./geocoders/here":5,"./geocoders/mapbox":6,"./geocoders/mapquest":7,"./geocoders/mapzen":8,"./geocoders/nominatim":9,"./geocoders/photon":10,"./geocoders/what3words":11}],13:[function(_dereq_,module,exports){
 (function (global){
 var L = (typeof window !== "undefined" ? window['L'] : typeof global !== "undefined" ? global['L'] : null),
 	lastCallbackId = 0,
@@ -1267,4 +1267,5 @@ module.exports = {
 };
 
 }).call(this,typeof global !== "undefined" ? global : typeof self !== "undefined" ? self : typeof window !== "undefined" ? window : {})
-},{}]},{},[12]);
+},{}]},{},[12])(12)
+});