+++ web/assets/js/site.js
@@ -1,26 +1,28 @@
+// debug mode
+var siteDebug = false;
 // personal api keys
 var mapboxKey = 'pk.eyJ1IjoibWFwYm94IiwiYSI6ImNpbG10dnA3NzY3OTZ0dmtwejN2ZnUycjYifQ.1W5oTOnWXQ9R1w8u3Oo1yA';
 var thuforKey = '4fc2613fe5d34ca697a03ea1dc8f0b2b';
 // overpass layer options
 var minOpZoom = 15, email = 'info@bexhill-osm.org.uk';
-// map bounds
-var mapLeft = 0.3000, mapTop = 50.8200, mapRight = 0.5350, mapBottom = 50.8800;
-// map centre
-var mapLat = 50.8424, mapLng = 0.4676, mapZoom = 15;
+// map area
+var mapBounds = { lef: 0.3000, top: 50.8200, rig: 0.5350, bot: 50.8800 };
+// map open location
+var mapCentre = [50.8424, 0.4676], mapZoom = 15;
 // map base layer
 var actTileLayer = 'osmstd';
 // tab to open
 var actTab = 'home';
 
 // swipe away sidebar on larger touch devices
-$(function() {
-	$('.sidebar-content').on('swipeleft', function() {
+$(function () {
+	$('.sidebar-content').on('swipeleft', function () {
 		if ($(window).width() >= 768 && (L.Browser.touch)) sidebar.close();
 	});
 });
 
 // smooth scrolling to poi anchor
-$(document).on('click', 'a[href*="#goto"]', function(e) {
+$(document).on('click', 'a[href*="#goto"]', function (e) {
 	if (location.pathname.replace(/^\//, '') === this.pathname.replace(/^\//, '') && location.hostname === this.hostname) {
 		var target = $(this.hash);
 		target = target.length ? target : $('[id=' + this.hash.slice(1) + ']');
@@ -31,27 +33,24 @@ $(document).on('click', 'a[href*="#goto"]', function(e) {
 	}
 });
 
-$('.sidebar-tabs').click(function() {
+$('.sidebar-tabs').click(function () {
 	// clear checkboxes if moving to another poi tab
 	if (actTab !== $('.sidebar-pane.active').attr('id')) {
 		actTab = $('.sidebar-pane.active').attr('id');
 		if (actTab === 'shops' || actTab === 'amenities' || actTab === 'services' || actTab === 'leisure') clear_map();
 	}
 	// resize links on mini map
-	if (actTab === 'home') setTimeout(function() { $('map').imageMapResize(); }, 500);
+	if (actTab === 'home') setTimeout(function () { $('map').imageMapResize(); }, 500);
 });
 
-// don't escape HTML string in Mustache
-Mustache.escape = function(text) { return text; };
-
 // https://github.com/davidjbradshaw/image-map-resizer
-$(document).ready(function() {
+$(document).ready(function () {
 	// add small delay after load for sidebar to animate open
-	setTimeout(function() {	$('map').imageMapResize(); }, 500);
+	setTimeout(function () { $('map').imageMapResize(); }, 500);
 	// set collapsible accordion for opening hours
-	map.on('popupopen', function() {
+	map.on('popupopen', function () {
 		// delay required when switching popups immediately
-		setTimeout(function() {
+		setTimeout(function () {
 			$('#accordOh').accordion({
 				heightStyle: 'content',
 				collapsible: true,
@@ -65,11 +64,9 @@ $(document).ready(function() {
 // https://github.com/aratcliffe/Leaflet.contextmenu
 var map = new L.map('map', {
 	contextmenu: true,
-	contextmenuWidth: 120,
 	contextmenuItems: [{
 		text: '<i class="fa fa-search"></i>&nbsp; Lookup',
 		callback: reverseLookup,
-		index: 0
 	}, '-', {
 		text: '<i class="fa fa-map-marker"></i>&nbsp; Add walk point',
 		callback: walkPoint
@@ -82,12 +79,13 @@ var geoMarker, rLookup = false;
 function reverseLookup(e) {
 	// get location, look up id on nominatim and pass it to overpass
 	var geocoder = L.Control.Geocoder.nominatim();
-	geocoder.reverse(e.latlng, map.options.crs.scale(map.getZoom()), function(results) {
+	geocoder.reverse(e.latlng, map.options.crs.scale(map.getZoom()), function (results) {
 		geoMarker = results[0];
 		if (geoMarker) {
+			if (siteDebug) console.debug(geoMarker);
 			if (markerId) clear_map();
 			rLookup = true;
-			show_overpass_layer(geoMarker.properties.osm_type + '(' + geoMarker.properties.osm_id + ');');
+			show_overpass_layer(geoMarker.properties.osm_type + '(' + geoMarker.properties.osm_id + ')(' + [mapBounds.top, mapBounds.lef, mapBounds.bot, mapBounds.rig] + ');');
 		}
 	});
 }
@@ -95,13 +93,13 @@ function walkPoint(e) {
 	if ($(window).width() >= 768) sidebar.open('walking');
 	// drop a walk marker if one doesn't exist
 	var wp = routingControl.getWaypoints();
-	for(var i = 0; i < wp.length; i++){
-		if(!wp[i].name){
-			routingControl.spliceWaypoints(i, 1, e.latlng);
+	for (var c = 0; c < wp.length; c++){
+		if(!wp[c].name){
+			routingControl.spliceWaypoints(c, 1, e.latlng);
 			return;
 		}
 	}
-    routingControl.spliceWaypoints(wp.length, 0, e.latlng);
+	routingControl.spliceWaypoints(wp.length, 0, e.latlng);
 }
 function improveMap(e) {
 	// create a note on osm.org
@@ -119,10 +117,9 @@ var tileLayerData = {
 		url: 'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
 		zoom: 19
 	},
-	osmfr: {
-		name: 'OpenStreetMap (france)',
-		url: 'https://{s}.tile.openstreetmap.fr/osmfr/{z}/{x}/{y}.png',
-		attribution: '<a href="http://openstreetmap.fr/" target="_blank">OSM France</a>',
+	osmhot: {
+		name: 'OpenStreetMap (humanitarian)',
+		url: 'https://{s}.tile.openstreetmap.fr/hot/{z}/{x}/{y}.png',
 		zoom: 19
 	},
 	cycle: {
@@ -172,11 +169,9 @@ var attribution = '&copy; <a href="http://openstreetmap.org/copyright" title="Co
 var tileLayers = {};
 var tileList = {name: [], keyname: []};
 for (var tile in tileLayerData) {
-	var tileAttribution;
 	var tilemaxZoom = tileLayerData[tile].zoom;
 	var subdomains = tileLayerData[tile].subdomains ? tileLayerData[tile].subdomains : 'abc';
-	if (tileLayerData[tile].attribution) tileAttribution = tileLayerData[tile].attribution + ' | ' + attribution;
-	else tileAttribution = attribution;
+	var tileAttribution = (tileLayerData[tile].attribution) ? tileLayerData[tile].attribution + ' | ' + attribution : attribution;
 	tileLayers[tileLayerData[tile].name] = L.tileLayer(
 		tileLayerData[tile].url, { maxNativeZoom: tilemaxZoom, maxZoom: 20, attribution: tileAttribution, subdomains: subdomains }
 	);
@@ -188,21 +183,15 @@ L.control.layers(tileLayers).addTo(map);
 L.control.scale({ metric: false, position: 'bottomright' }).addTo(map);
 
 // get basemap layer name on change
-map.on('baselayerchange', function(e) {
-  for (var x = 0; x < tileList.name.length; x++) {
-	if (e.name === tileList.name[x]) actTileLayer = tileList.keyname[x];
-  }
-});
-
-// reverselookup only on high zoom level
-map.on('zoomend', function() {
-	if (map.getZoom() >= 18) map.contextmenu.setDisabled(0, false);
-	else map.contextmenu.setDisabled(0, true);
+map.on('baselayerchange', function (e) {
+	for (var c = 0; c < tileList.name.length; c++) {
+		if (e.name === tileList.name[c]) actTileLayer = tileList.keyname[c];
+	}
 });
 
 // middle-mouse button reverselookup
-map.on('mouseup', function(e) {
-	if (e.originalEvent.button === 1 && map.getZoom() >= 18) reverseLookup(e);
+map.on('mouseup', function (e) {
+	if (e.originalEvent.button === 1) reverseLookup(e);
 });
 
 // https://github.com/domoritz/leaflet-locatecontrol
@@ -216,7 +205,7 @@ L.control.locate({
 		popup: 'You are within {distance} {unit} from this point',
 		outsideMapBoundsMsg: 'You appear to be located outside Bexhill.  Come visit!'
 	},
-	onLocationError: function() {
+	onLocationError: function () {
 		alert('Sorry, there was an error while trying to locate you.');
 	}
 }).addTo(map);
@@ -226,8 +215,7 @@ L.Control.geocoder({
 	geocoder: L.Control.Geocoder.nominatim({
 		geocodingQueryParams: {
 			bounded: 1,
-			viewbox: mapLeft + ',' + mapTop + ',' + mapRight + ',' + mapBottom,
-			extratags: 1,
+			viewbox: [mapBounds.lef, mapBounds.top, mapBounds.rig, mapBounds.bot],
 			email: email
 		}
 	}),
@@ -236,19 +224,24 @@ L.Control.geocoder({
 	position: 'topleft',
 	placeholder: 'Type address or place name...'
 })
-.on('markgeocode', function(e) {
+.on('markgeocode', function (e) {
 	// pass nominatim address lookup to overpass
 	if (markerId) clear_map();
 	rLookup = true;
 	geoMarker = e.geocode;
+	if (siteDebug) console.debug(geoMarker);
 	if (map.getZoom() <= minOpZoom) map.setZoom(minOpZoom, { animate: false });
-	map.flyTo(geoMarker.center);
-	show_overpass_layer(geoMarker.properties.osm_type + '(' + geoMarker.properties.osm_id + ');');
+	show_overpass_layer(geoMarker.properties.osm_type + '(' + geoMarker.properties.osm_id + ')(' + [mapBounds.top, mapBounds.lef, mapBounds.bot, mapBounds.rig] + ');');
 	$(':input','.leaflet-control-geocoder-form').blur();
 	$(':input','.leaflet-control-geocoder-form').val('');
 })
 .addTo(map);
 $('.leaflet-control-geocoder-icon').attr('title','Find address');
+
+function geocoderLink() {
+	if ($(window).width() < 768) sidebar.close();
+	$('.leaflet-control-geocoder-icon').click();
+}
 
 // https://github.com/cliffcloud/Leaflet.EasyButton
 // permalink button
@@ -257,7 +250,7 @@ L.easyButton({
 	states:[{
 		icon:'fa fa-link',
 		title:'Get map link',
-		onClick:function() {
+		onClick:function () {
 			var uri = URI(window.location.href);
 			var selectedPois = [], walkCoords = [], openNow;
 			var walkWayp = routingControl.getWaypoints();
@@ -267,7 +260,7 @@ L.easyButton({
 				}
 			}
 			if ($('input.opennow').is(':checked')) openNow = 1;
-			$('#pois' + actTab + ' input.poi-checkbox:checked').each(function(i, element) { selectedPois.push(element.dataset.key);	});
+			$('#pois' + actTab + ' input.poi-checkbox:checked').each(function (i, element) { selectedPois.push(element.dataset.key);	});
 			// M = basemap, T = tab, O = opennow, P = grouped pois, I = single poi, W = walkpoints
 			uri.query({ 'M': actTileLayer, 'T': actTab, 'O': openNow, 'P': selectedPois, 'I': markerId, 'W': walkCoords });
 			window.prompt('Copy this link to share current map:', uri.href());
@@ -280,7 +273,7 @@ L.easyButton({
 	states:[{
 		icon:'fa fa-trash',
 		title:'Clear map',
-		onClick:function() {
+		onClick:function () {
 			// clear walk points, roll up suggested walks, clear poi marker layers
 			routingControl.setWaypoints([]);
 			$('#accordWk').accordion({ active: false });
@@ -289,7 +282,7 @@ L.easyButton({
 	}]
 }).addTo(map);
 // full reload on right click
-$('#btnclearmap').bind('contextmenu',function() {
+$('#btnclearmap').bind('contextmenu',function () {
 	$(location).attr('href', window.location.pathname);
 	return false;
 });
@@ -321,7 +314,7 @@ var routingControl = L.Routing.control({
 	geocoder: L.Control.Geocoder.nominatim({
 		geocodingQueryParams: {
 			bounded: 1,
-			viewbox: mapLeft + ',' + mapTop + ',' + mapRight + ',' + mapBottom
+			viewbox: [mapBounds.lef, mapBounds.top, mapBounds.rig, mapBounds.bot]
 		}
 	})
 });
@@ -335,66 +328,63 @@ $('#accordWk').accordion({
 });
 
 // https://github.com/pawelczak/EasyAutocomplete
-var poitags = [];
-var category = [];
-var i = 0;
-// get all tags
+var poitags = [], category = [], c = 0;
+// get all keywords
 for (var poi in pois) {
 	poitags += '"' + pois[poi].tabName + '~' + poi + '": ' + JSON.stringify(pois[poi].tagKeyword) + ', ';
-	category[i] = { listLocation: pois[poi].tabName + '~' + poi, header: pois[poi].tabName + ' - ' + pois[poi].name };
-	i++;
+	category[c] = { listLocation: pois[poi].tabName + '~' + poi, header: pois[poi].tabName + ' - ' + pois[poi].name };
+	c++;
 }
 // clean up and covert to array
 poitags = poitags.substring(0, poitags.length - 2);
-poitags = JSON.parse('{ ' + poitags + '}');
+poitags = JSON.parse('{' + poitags + '}');
 // options for autocomplete
 var options = {
 	data: poitags,
 	minCharNumber: 3,
 	list: {
 		maxNumberOfElements: 10,
-		onChooseEvent: function() {
+		match: { enabled: true },
+		onChooseEvent: function () {
 			// Find selected items category, split it to get checkbox, then display
 			var z = ($('#autocomplete').getSelectedItemIndex());
-			var catsplit = (document.getElementsByClassName('eac-category')[z].innerText);
-			catsplit = catsplit.split(' - ');
-			actTab = catsplit[0];
+			var catSplit = (document.getElementsByClassName('eac-category')[z].innerText);
+			catSplit = catSplit.split(' - ');
+			actTab = catSplit[0];
 			sidebar.open(actTab);
-			$('#pois' + actTab + ' input[id="' + catsplit[1] + '"]').prop('checked', true);
+			$('#pois' + actTab + ' input[id="' + catSplit[1] + '"]').prop('checked', true);
 			// Highlight checkbox or hide sidebar for mobile users
-			if ($(window).width() >= 768) {
-				$('#pois' + actTab + ' input[id="' + catsplit[1] + '"]').parent().parent().parent().effect('highlight', {}, 3000);
-			}
-			else sidebar.close();
+			if ($(window).width() >= 768) $('#pois' + actTab + ' input[id="' + catSplit[1] + '"]').parent().parent().parent().effect('highlight', {}, 3000);
 			setting_changed();
 			$('#autocomplete').val('');
-	},
-	match: {
-		enabled: true
 		}
 	},
-	categories: [
-		category[0]
-	]
+	categories: [category[0]]
 };
 // push categories into options array
-for (var x = 1; x < category.length; x++) { options.categories.push(category[x]); }
+for (var c = 1; c < category.length; c++) { options.categories.push(category[c]); }
 $('#autocomplete').easyAutocomplete(options);
 $('div.easy-autocomplete').removeAttr('style');
 
 // https://github.com/kartenkarsten/leaflet-layer-overpass
 var spinner = 0, markerId;
 function callback(data) {
+	var type, markerPopup;
 	// set marker popup dimensions on screensize
-	var type, markerPopup, customOptions = ($(window).width() > 500) ? { maxWidth: 350, minWidth: 205 } : { maxWidth: 250, minWidth: 205, autoPanPaddingTopLeft: 30, className: 'popup-mobile' };
-	if (spinner > 0) spinner -= 1;
+	var customOptions = ($(window).width() >= 768) ? { maxWidth: 350 } : { maxWidth: 250, className: 'popup-mobile' };
+	customOptions.autoPanPaddingTopLeft = (($(window).width() >= 768) && rLookup) ? [sidebar.width()+50,5] : [30,5];
+	customOptions.autoPanPaddingBottomRight = [5,75];
+	customOptions.minWidth = 205;
+	customOptions.closeButton = false;
+	if (spinner > 0) spinner--;
 	if (spinner === 0) $('#spinner').hide();
-	for(i = 0; i < data.elements.length; i++) {
-		var e = data.elements[i];
+	for (var c = 0; c < data.elements.length; c++) {
+		var e = data.elements[c];
 		if (e.id in this.instance._ids) continue;
 		this.instance._ids[e.id] = true;
 		var pos = (e.type === 'node') ? new L.LatLng(e.lat, e.lon) : new L.LatLng(e.center.lat, e.center.lon);
 		type = undefined;
+		if (siteDebug) console.debug(e);
 		if (e.tags.amenity) {
 			if (!type) type = e.tags.amenity;
 			// Group similar pois
@@ -463,12 +453,12 @@ function callback(data) {
 		}
 		var poi = pois[type];
 		var iconName = poi ? poi.iconName : '000blank';
-		var markerIcon  = L.icon({
+		var markerIcon = L.icon({
 			iconUrl: 'assets/img/icons/' + iconName + '.png',
 			iconSize: [32, 37],
-			iconAnchor: [18.5, 35],
+			iconAnchor: [16, 35],
 			shadowUrl: 'assets/img/icons/000shadow.png',
-			shadowAnchor: [17, 27],
+			shadowAnchor: [16, 27],
 			popupAnchor: [0, -27]
 		});
 		var marker = L.marker(pos, {
@@ -527,9 +517,10 @@ function setting_changed(newcheckbox) {
 		iconLayer.clearLayers();
 		markerId = undefined;
 		if ($('input.poi-checkbox:checked').length > 0) {
+			if ($(window).width() < 768) sidebar.close();
 			//build overpass query
 			var query = '(';
-			$('#pois' + actTab + ' input.poi-checkbox:checked').each(function(i, element) {
+			$('#pois' + actTab + ' input.poi-checkbox:checked').each(function (i, element) {
 				query += 'node' + pois[element.dataset.key].query + '(BBOX);';
 				query += 'way' + pois[element.dataset.key].query + '(BBOX);';
 			});
@@ -546,26 +537,25 @@ function setting_changed(newcheckbox) {
 
 // build content for sidebar pois
 function show_pois_checkboxes(tabName) {
-	var i = 0;
+	var c = 0;
 	var content = '<div class="anchor"><a id="gototop' + tabName + '" href="#gotobot' + tabName + '">| <i class="fa fa-arrow-down"></i> |</a></div>';
 	content += '<table style="width:100%;">';
 	for (var poi in pois) {
 		if (pois[poi].tabName === tabName) {
-			if (i % 2 === 0) content += '<tr>';
+			if (c % 2 === 0) content += '<tr>';
 			content += '<td style="overflow:hidden; white-space:nowrap;">';
-			var checkbox = Mustache.render(
+			content += L.Util.template(
 				'<div class="poi-checkbox"> \
-					<label title="{{name}}"> \
-						<img style="width:28px;" src="assets/img/icons/{{icon}}.png"></img> \
-						<input type="checkbox" class="poi-checkbox" id="{{name}}" data-key="{{key}}" onclick="setting_changed(&#39;{{key}}&#39;)"><span>{{name}}</span> \
+					<label title="{name}"> \
+						<img style="width:28px;" src="assets/img/icons/{icon}.png"></img> \
+						<input type="checkbox" class="poi-checkbox" id="{name}" data-key="{key}" onclick="setting_changed(&#39;{key}&#39;)"><span>{name}</span> \
 					</label> \
 				</div>',
 				{ key: poi, name: pois[poi].name, icon: pois[poi].iconName, tabname: pois[poi].tabName }
 			);
-			content += checkbox;
 			content += '</td>';
-			i++;
-			if (i % 2 === 0) content += '</tr>';
+			c++;
+			if (c % 2 === 0) content += '</tr>';
 		}
 	}
 	content += '</table>';
@@ -579,9 +569,9 @@ show_pois_checkboxes('leisure');
 show_pois_checkboxes('tourism');
 
 // suggested walk waypoints
-function suggwalk(walkname) {
+function suggWalk(walkName) {
 	clear_map();
-	if (walkname === 'tmt') {
+	if (walkName === 'tmt') {
 		routingControl.setWaypoints([
 			[50.84059, 0.49121],
 			[50.83729, 0.47612],
@@ -592,15 +582,15 @@ function suggwalk(walkname) {
 		show_overpass_layer('(node["ref"~"^TMT"](BBOX));');
 		map.flyTo([50.8385, 0.4787], 16);
 	}
-	else if (walkname === '1066') {
+	else if (walkName === '1066') {
 		routingControl.setWaypoints([
 			[50.84522, 0.48044],
 			[50.84972, 0.48419],
 			[50.87800, 0.50009]
 		]);
-		map.flyTo([50.8551, 0.4527], 14);
+		map.flyTo([50.8504, 0.4840], 15);
 	}
-	else if (walkname === 'bc1') {
+	else if (walkName === 'bc1') {
 		routingControl.setWaypoints([
 			[50.84125, 0.47717],
 			[50.84515, 0.47961],
@@ -610,9 +600,9 @@ function suggwalk(walkname) {
 			[50.83792, 0.47660],
 			[50.84093, 0.47718]
 		]);
-		map.flyTo([50.8526, 0.4526], 14);
+		map.flyTo([50.8474, 0.4874], 14);
 	}
-	else if (walkname === 'wwh') {
+	else if (walkName === 'wwh') {
 		routingControl.setWaypoints([
 			[50.83567, 0.45892],
 			[50.83701, 0.47363]
@@ -628,25 +618,32 @@ if (uri.hasQuery('T')) actTab = uri.search(true).T;
 if (uri.hasQuery('O')) $('input.opennow').prop('checked', uri.search(true).O);
 if (uri.hasQuery('W')) {
 	var walkCoords = uri.search(true).W;
-	for (i = 0; i < walkCoords.length; i++) {
-		walkCoords[i] = walkCoords[i].replace('~', ', ');
-		routingControl.spliceWaypoints(i, 1, JSON.parse('[' + walkCoords[i] + ']'));
+	for (var c = 0; c < walkCoords.length; c++) {
+		walkCoords[c] = walkCoords[c].replace('~', ', ');
+		routingControl.spliceWaypoints(c, 1, JSON.parse('[' + walkCoords[c] + ']'));
 	}
 }
 if (uri.hasQuery('P')) {
-	var poi = uri.search(true).P;
-	if (!$.isArray(poi)) {
-		$('#pois' + uri.search(true).T + ' input[data-key=' + poi + ']').prop('checked', true);
-	}
-	else {
-		for (i = 0; i < poi.length; i++) {
-			$('#pois' + uri.search(true).T + ' input[data-key=' + poi + ']').prop('checked', true);
+	var selectedPois = uri.search(true).P;
+	if (!$.isArray(selectedPois)) {
+		$('#pois' + uri.search(true).T + ' input[data-key=' + selectedPois + ']').prop('checked', true);
+		// highlight checkbox or hide sidebar for mobile users
+		if ($(window).width() >= 768) {
+			sidebar.open(actTab);
+			$('#pois' + uri.search(true).T + ' input[data-key=' + selectedPois + ']').parent().parent().parent().effect('highlight', {}, 3000);
 		}
 	}
-	// highlight checkbox or hide sidebar for mobile users
-	if ($(window).width() >= 768) {
-		sidebar.open(actTab);
-		$('#pois' + uri.search(true).T + ' input[data-key=' + poi + ']').parent().parent().parent().effect('highlight', {}, 3000);
+	else {
+		for (var c = 0; c < selectedPois.length; c++) {
+			// the last poi has a "/" on it because leaflet-hash
+			var multiplePois = selectedPois[c].replace('/', '');
+			$('#pois' + actTab + ' input[data-key=' + multiplePois + ']').prop('checked', true);
+			// highlight checkbox or hide sidebar for mobile users
+			if ($(window).width() >= 768) {
+				sidebar.open(actTab);
+				$('#pois' + actTab + ' input[data-key=' + multiplePois + ']').parent().parent().parent().effect('highlight', {}, 3000);
+			}
+		}
 	}
 	setting_changed();
 }
@@ -654,23 +651,24 @@ if (uri.hasQuery('I')) {
 	rLookup = true;
 	markerId = uri.search(true).I;
 	var splitId = markerId.split('~');
-	show_overpass_layer(splitId[0] + '(' + splitId[1] + ');');
+	show_overpass_layer(splitId[0] + '(' + splitId[1] + ')(' + [mapBounds.top, mapBounds.lef, mapBounds.bot, mapBounds.rig] + ');');
 }
 // if not returning from a permalink, give defaults
 if (!uri.hasQuery('W') || !uri.hasQuery('P') || !uri.hasQuery('I')) {
-	if (window.location.hash.indexOf('/') !== 3) map.setView([mapLat, mapLng], mapZoom);
+	if (window.location.hash.indexOf('/') !== 3) map.setView(mapCentre, mapZoom);
 	if ($(window).width() >= 768) sidebar.open(actTab);
 }
 tileLayers[tileLayerData[actTileLayer].name].addTo(map);
-map.setMaxBounds([[mapTop, mapLeft], [mapBottom, mapRight]]);
+map.setMaxBounds([[mapBounds.top, mapBounds.lef], [mapBounds.bot, mapBounds.rig]]);
 	
 function show_overpass_layer(query) {
+	if (siteDebug) console.debug(query);
 	if (!query || query === '();') {
 		console.log('There is nothing selected to filter by.');
 		return;
 	}
 	var opl = new L.OverPassLayer({
-		debug: false,
+		debug: siteDebug,
 		minzoom: minOpZoom,
 		query: query + 'out center;&contact=' + email, //contact info only for use with .fr endpoint
 		endpoint: 'https://api.openstreetmap.fr/oapi/interpreter/',